refactor: modularize HomePageV2 UI into dedicated widgets for header, search, and surah list
